requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
# psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
WeasyPrint==52.4
markdown2==2.4.10
gunicorn==21.2.0
httpx==0.24.1
Pillow==10.0.0